Change AuthenticationManager to class with lock
# Motivation
Currently, the JWT authentication is using two actors to isolate the state. This leads to thread hops for every APNS request which we can avoid with using a `Lock` instead.

# Modification
This PR contains a couple of modifications to the JWT authentication handling:
- Refactor the two actors to a single class using a `Lock`
- Drop the dependency on `Foundation` from the JWT handling.
- Improve the logging
- Add tests to verify the behaviour of the new JWT generation.

# Result
We now have a faster and better tested JWT handling.