Refactor configuration and client
# Motivation
This PR is doing multiple things in once and I can't split them apart since the client and the configuration are tightly coupled. The goals of this PR is to create an evolable API that we can use to build more semantically safe constructs. Furthermore, it splits the configuration properly between what can be configured on a single connection and what can be configured on a request.

# Modification
In details this PR does the following things:

1. Every type that is touched in this PR should be forward evolable without API breakage. Therefore, all enums are converted to struct
2. The current configuration mixes configs from the connection and single requests. The new proposed config splits them between the two types.
3. Introduce a new error struct that contains semantic information and more debug information
4. Makes the client generic over the encoder/decoder and introduces a new low level `send` method.

# Result
We now have a new config and a baseline client that allows us to build semantically safe APIs in the next PRs.